Students panel is now toggleable. Meaning the perfoamnce impact is minimal
