+++ EduAR/Assets/Scripts/StudentListPopUp.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 916ce6c48e43e664f94145ecbc1298d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
